+++ gamestore/src/main/java/com/company/gamestore/repositories/GameRepo.java
@@ -3,6 +3,6 @@ package com.company.gamestore.repositories;
 import com.company.gamestore.models.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GameRepo extends JpaRepository<Game, Long> {
+public interface GameRepo extends JpaRepository<Game, Integer> {
     // Add custom query methods here if needed
 }
